src/third_party/blink/web_tests/virtual/controls-refresh/datetimelocal-picker/datetimelocal-month-switcher-buttons-invalid.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../fast/forms/resources/common.js"></script>
<script src="../../../fast/forms/resources/picker-common.js"></script>
<script src="../../../fast/forms/calendar-picker/resources/calendar-picker-common.js"></script>
</head>
<body>
<input type="datetime-local" id="datetime-local-0" value="2019-05-22T13:02" max="2019-06-15T13:02">
<input type="datetime-local" id="datetime-local-1" value="2019-11-20T13:02" min="2019-10-29T13:02">
<input type="datetime-local" id="datetime-local-2" value="2019-03-15T13:02" step="604800">
<input type="datetime-local" id="datetime-local-3" value="2019-03-15T13:02" step="691200">
<input type="datetime-local" id="datetime-local-4" value="2019-08-02T13:02" step="604800">
<script>

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-0');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickNextMonthButton();
    assert_equals(dateTimeElement.value, "2019-06-15T13:02", "Month chooser button should have changed month and stopped at max date");

    clickDayCellAt(2, 2);
    eventSender.keyDown('Enter');
    assert_equals(dateTimeElement.value, "2019-06-11T13:02", "Clicking date should have changed date");
    assert_equals(internals.pagePopupWindow, null, "Popup should have closed");
  });
}, "Datetimelocal picker: Next month button should advance to next month and clip selection to max date.");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-1');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickPrevMonthButton();
    assert_equals(dateTimeElement.value, "2019-10-29T13:02", "Month chooser button should have changed month and stopped at min date");

    clickDayCellAt(4, 4);
    eventSender.keyDown('Enter');
    assert_equals(dateTimeElement.value, "2019-10-31T13:02", "Clicking date should have changed date");
    assert_equals(internals.pagePopupWindow, null, "Popup should have closed");
  });
}, "Datetimelocal picker: Previous month button should advance to previous month and clip selection to min date.");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-2');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickNextMonthButton();
    assert_equals(dateTimeElement.value, "2019-04-12T13:02", "Closest valid date to 4-15 should have been selected");

    clickDayCellAt(5, 0);
    eventSender.keyDown('Enter');
    assert_equals(dateTimeElement.value, "2019-04-05T13:02", "Clicking date should have changed date");
    assert_equals(internals.pagePopupWindow, null, "Popup should have closed");
  });
}, "Datetimelocal picker: If same date in next month is invalid, next month button should choose closest value -- smaller date case");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-3');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickNextMonthButton();
    assert_equals(dateTimeElement.value, "2019-04-16T13:02", "Closest valid date to 4-15 should have been selected");

    clickDayCellAt(3, 3);
    eventSender.keyDown('Enter');
    assert_equals(dateTimeElement.value, "2019-04-24T13:02", "Clicking date should have changed date");
    assert_equals(internals.pagePopupWindow, null, "Popup should have closed");
  });
}, "Datetimelocal picker: If same date in next month is invalid, next month button should choose closest value -- larger date case");

promise_test(() => {
  let dateTimeElement = document.getElementById('datetime-local-4');
  return openPickerWithPromise(dateTimeElement)
  .then(() => {
    // Make the picker dismiss synchronously so we don't need to insert
    // an artificial delay in the test
    internals.pagePopupWindow.CalendarPicker.commitDelayMs = 0;

    clickNextMonthButton();
    assert_equals(dateTimeElement.value, "2019-09-06T13:02", "Date in new month should be selected even though previous valid date is closer");

    clickDayCellAt(5, 2);
    eventSender.keyDown('Enter');
    assert_equals(dateTimeElement.value, "2019-09-20T13:02", "Clicking date should have changed date");
    assert_equals(internals.pagePopupWindow, null, "Popup should have closed");
  });
}, "Datetimelocal picker: If closest target date is back in the current month, still jump to the value in the next month");

</script>
</body>
</html>
